refactor: split EotE/Genesys schemas into separate folders
Reorganizes the schema structure for better separation of concerns:
- shared/ - common character and supplement schemas
- eote/ - Edge of the Empire specific schemas
- genesys/ - Genesys specific schemas
